% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ciPlot}
\alias{ciPlot}
\title{Create a confidence interval plot}
\usage{
ciPlot(
  data,
  col = 1,
  append = FALSE,
  ylim = c(min(as.matrix(data$low)[, col]), max(as.matrix(data$high)[, col])),
  ci_col = "lightgrey",
  ...
)
}
\arguments{
\item{data}{The data to plot, should be a list that contains "mean", "low" and "high" matrices}

\item{col}{Which column of each matrix contains the data to plot, defaults to 1}

\item{append}{Should we create a new plot or append a confidence interval to one already active, defaults to FALSE}

\item{ylim}{The y axis limits for the plot, defaults to min of low and max of high}

\item{ci_col}{The confidence area color, defaults to lightgrey}

\item{...}{Additional arguments to pass to the plot function.}
}
\description{
Create a confidence interval plot
}
