% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{createRandDir}
\alias{createRandDir}
\title{Create a folder with a random name inside the current folder}
\usage{
createRandDir()
}
\value{
The name of the new folder which is guaranteed to be unique
}
\description{
Create a folder with a random name inside the current folder
}
