% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{JAnimateGIF}
\alias{JAnimateGIF}
\title{Combine multiple plots into an animated GIF file.}
\usage{
JAnimateGIF(nFrames = NULL, frameKeys = 1:nFrames, gifFileName, plotFn,
  frameRate = 30, tmpDir = tempdir(TRUE), ...)
}
\arguments{
\item{nFrames}{Number of frames to be generated. You must specify one of
\code{nFrames} or \code{frameKeys}.}

\item{frameKeys}{Vector of keys to be passed to `plotFn` to identify the
frame to be plotted. If not specified, `frameKeys` will be set to the
sequence `1:nFrames`.}

\item{gifFileName}{Name of the GIF file to be created.}

\item{plotFn}{Function which is called once for each frame. It is called once
for each frame to be generated, with a single argument which is one of the
values from \code{frameKeys}. If it does not generate a plot, the frame
will be silently skipped.}

\item{frameRate}{Play back frame rate - used to set the frame delay in the
GIF file.}

\item{...}{Any additional arguments are passed to the \code{JPlotToPNG}
function.}

\item{tmpdir}{Name of a directory to be used to create temporary files in.}
}
\value{
The error message (as a character vector) from the ImageMack convert
  command, or \code{character(0)} on success.
}
\description{
Works by writing each plot to a PNG file, then combining them into a GIF file
using the ImageMagick utility. Frames are created by calling a custom plot
function.
}
\details{
You must have \href{https://imagemagick.org/script/download.php}{ImageMagick}
installed, and the ImageMagick bin subdirectory must be in your PATH
environment variable. If ImageMagick is not installed, or is not in your
PATH, \code{JAnimateGIF} will fail with an exception such as:

\preformatted{Error in system2("magick", c("convert", "jp*.png", "-delay",
100/frameRate, : '"xmagick"' not found}.

For a general discussion about creating an animation in R, see
\url{https://stackoverflow.com/questions/1298100/creating-a-movie-from-a-series-of-plots-in-r},
or \url{https://www.r-graph-gallery.com/animation/}.

The png to gif conversion can be quite time consuming for a large number of
frames, so start with a small number of frames to ensure everything works as
expected.
}
\examples{
nFrames <- 50
.plotPoly <- function(frame) {
  plot(NULL, xlim = c(-1, 1), ylim = c(-1, 1), asp = 1, xlab = "", ylab = "", axes = FALSE)
  # Frame starts from 1, but a 1 cornered polygon is not very visually interesting
  nCorners <- frame + 1
  angles <- 2 * pi * (1:nCorners) / nCorners
  f <- frame / nFrames
  col <- rgb(sin(pi / 2 * f), cos(pi / 2 * f), sin(pi / 2 * (f + .5)))
  polygon(cos(angles), sin(angles), col = col, border = "black", lwd = 4)
}

JAnimateGIF(nFrames, gifFileName = "../poly.gif", plotFn = .plotPoly, frameRate = 10)

}
