% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim-tools.R
\name{JAnimateScenes}
\alias{JAnimateScenes}
\title{Animate a list of \code{\link{Jscene}}s.}
\usage{
JAnimateScenes(gifFileName, scenes, ...)
}
\arguments{
\item{gifFileName}{Name of the GIF file to be created.}

\item{scenes}{A list of scene objects, each one created by calling the
\code{\link{JScene}} constructor.}

\item{...}{Additional poarameters are passed on to \code{\link{JAnimateGIF}}.}
}
\value{
The list of arguments passed to the \code{\link{JAnimateGIF}} function (invisibly).
}
\description{
Animate a list of \code{\link{Jscene}}s.
}
\seealso{
\code{\link{JAnimateGIF}}, \code{\link{Jscenes}}
}
