% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{JAnimateUnwrapped}
\alias{JAnimateUnwrapped}
\title{Prepare to create frames for an animation}
\usage{
JAnimateUnwrapped(
  videoFileName,
  frameRate = 30,
  loop = 0,
  method = c("auto", "magick-r", "gifski", "magick", "ffmpeg"),
  optimize = FALSE,
  progress = FALSE,
  tmpDir = tempdir(TRUE)
)
}
\value{
A function that should be used to plot each frame. See examples
}
\description{
Similar to JAnimateGIF except when you want to create frames within your own loop.
}
\examples{
frameFn <- JAnimateUnwrapped("anim.gif")
nFrames <- 50
for (frame in 1:nFrames) {
  frameFn(plotExpr = {
    plot(NULL, xlim = c(-1, 1), ylim = c(-1, 1), asp = 1, xlab = "", ylab = "", axes = FALSE)
    # Frame starts from 1, but a 1 cornered polygon is not very visually interesting
    nCorners <- frame + 1
    angles <- 2 * pi * (1:nCorners) / nCorners
    f <- frame / nFrames
    # Pick a colour based on the frame number
    col <- rgb(sin(pi / 2 * f), cos(pi / 2 * f), sin(pi / 2 * (f + .5)))
    # Draw a polygon
    polygon(cos(angles), sin(angles), col = col, border = "black", lwd = 4)})
}

frameFn(complete = TRUE)

}
