% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{JCapSentence}
\alias{JCapSentence}
\title{Captialises the first letter in a string.}
\usage{
JCapSentence(s, strict = FALSE, collapse = NULL)
}
\arguments{
\item{s}{Character vector to be capitalised}

\item{strict}{If \code{TRUE}, the remainder of the string is converted to lower case.}

\item{collapse}{String used to collapse multiple character strings.}
}
\description{
Copied (and changed) from the doc for \code{toupper}.
}
\examples{
print(JCapSentence("the quick brown fox"))
# => [1] "The quick brown fox"

}
