% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{JCapWords}
\alias{JCapWords}
\title{Capitalises the first letter in each word in a string.}
\usage{
JCapWords(s, split = " ", sep = " ", ...)
}
\arguments{
\item{s}{Character vector to be capitalised.}

\item{split}{Regular expression used to split words to be capitalised.}

\item{sep}{String used to paste words back together.}

\item{...}{Additional arguments passed to \code{\link{JCapitalise}}}
}
\description{
Capitalises the first letter in each word in a string.
}
\examples{
print(JCapWords("the quick brown fox"))
# => [1] "The Quick Brown Fox"

}
