% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{JFill}
\alias{JFill}
\title{Fill "bad" values in a vector with the most recent "good" values}
\usage{
JFill(x, badValues = is.na(x), returnIndices = FALSE)
}
\arguments{
\item{x}{Vector to be filled}

\item{badValues}{Logical vector identifying the "bad" values in `x`, i.e. the
values to be replaced}

\item{returnIndices}{If TRUE, return indices into `x`, otherwise values from
`x`.}
}
\value{
description Either a vector of indices into `x` (when
  `returnIndices` is `TRUE`), or `x` with bad values replaced by the most
  recent good value.
}
\description{
Fill "bad" values in a vector with the most recent "good" values
}
\examples{
JFill(c(NA, 1, 2, 3, NA, 4, NA, NA))
## [1] NA  1  2  3  3  4  4  4

}
