% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_densities.R
\name{JPlotDensities}
\alias{JPlotDensities}
\title{Plots a list of densities as optionally filled lines using base graphics}
\usage{
JPlotDensities(densities, lineColours = NULL, fillColours = NULL,
  fillAlpha = NA, lty = 1, lwd = 2, xlim = NULL, ylim = NULL,
  add = FALSE, ...)
}
\arguments{
\item{densities}{List of densities to be plotted}

\item{lineColours}{Vector of colours used to draw the lines. Defaults to the
rainbow palette.}

\item{fillColours}{Vector of colours to fill the densitiy polygons. If not
specified or NULL, the polygons are not filled.}

\item{fillAlpha}{If not NA (the default), should be an alpha value between 0
and 1. The alpha of the fill colours are multiplied by this value.}

\item{lty}{Line style used to draw the lines.}

\item{lwd}{Line width used to draw the lines.}

\item{xlim}{Defines the graphical extents of the x-axis. Defaults to include all the density lines.}

\item{ylim}{Defines the graphical extents of the y-axis. Defaults to include all the density lines.}

\item{add}{If TRUE, the densities are added to an existing plot, otherwse a new plot is created.}

\item{...}{Additional parameters are passed to \code{\link[graphics]{plot}}.}
}
\description{
Attempts to provide sensible defaults while allowing maximum flexibility.
}
\examples{
data <- list(
  normal = rnorm(100),
  uniform = runif(50),
  exponential = rexp(200))
densities <- lapply(data, density)
JPlotDensities(densities)

}
\seealso{
\code{\link[stats]{density}}, \code{\link[RColorBrewer]{brewer.pal}}
}
