% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{JPlotRaster}
\alias{JPlotRaster}
\title{Adds a raster image to a plot, preserving the image's aspect ratio.}
\usage{
JPlotRaster(
  img,
  x,
  y,
  width = NA,
  height = NA,
  position = c("centre", "center", "bottomright", "bottom", "bottomleft", "left",
    "topleft", "top", "topright", "right"),
  ...
)
}
\arguments{
\item{img}{Raster image, possibly created by calling
\code{\link[jpeg]{readJPEG}} or \code{\link[png]{readPNG}}.}

\item{x, y}{Location on the plot of the image in user coordinates.}

\item{width}{Width of the image in user coordinates. The height will be
calculated from the aspect ratio of the image and the respective scales of
the x- and y-axes.}

\item{height}{Height of the image in user coordinates. The width will be
calculated from the aspect ratio of the image and the respective scales of
the x- and y-axes.}

\item{position}{Specifies the location of \code{x, y} on the image. E.g. if
\code{"centre"}, the image will be centred on \code{x, y}. if
\code{"topleft"}, the top left-hand corner of the image will be located at
\code{x, y}.}

\item{...}{Additional arguments are passed on to \code{\link[graphics]{rasterImage}}.}
}
\description{
Exactly one of \code{width} and \code{height} must be specified.
}
\examples{
\dontrun{
plot(...)
img <- readJPEG("myjpeg.jpg", native = TRUE)
# Draw image centred on (0, 0)
JPlotRaster(img, x = 0, y = 0, width = 2)
}

}
