% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim-tools.R
\name{JPlotScenes}
\alias{JPlotScenes}
\title{Construct a plotting function from a list of scenes.}
\usage{
JPlotScenes(scenes)
}
\arguments{
\item{scenes}{A list of scene objects, each one created by calling the
\code{\link{JScene}} constructor.}
}
\value{
A function that accepts a single argument, a frame number, and plots
  the appropriate frame from the animation. The function is intended to be
  used as the \code{plotFn} argument to \code{\link{JAnimateGIF}}.
}
\description{
\code{JPlotScenes} is not usually called directly, rather it is invoked from
inside \code{\link{JAnimateScenes}}. Combines a list of scenes, and returns a
function that plots a single frame from the appropriate scene.
}
\seealso{
\code{\link{JAnimateScenes}}
}
