% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_file.R
\name{JPlotToPDF}
\alias{JPlotToPDF}
\title{Plot to a PDF file}
\usage{
JPlotToPDF(
  filename,
  plotExpr,
  width = 180,
  height = NA,
  aspectRatio = 3/2,
  units = c("mm", "cm", "in"),
  bg = "white",
  paper = "special",
  family = "Helvetica",
  onlyIfDoesntExist = FALSE,
  createDirectory = TRUE,
  ...
)
}
\arguments{
\item{filename}{The name of the PDF to create or overwrite.}

\item{plotExpr}{A function or expression which will produce the plot to be
written to the file.}

\item{width}{The width of the output PDF file in \code{units}.}

\item{height}{The height of the output PDF file in \code{units}. Defaults to
\code{width / aspectRatio}. If \code{height} is specified,
\code{aspectRatio} will be ignored.}

\item{aspectRatio}{Aspect ratio (\code{width / height}) of the output PDF
file.}

\item{units}{Units of \code{width} and \code{height}.}

\item{bg}{Background colour - may be "transparent" for no background.}

\item{paper}{Paper size, defaults to "special" which is the value it must have if
\code{width} and \code{height} are to be used.}

\item{family}{The font family to be used.}

\item{onlyIfDoesntExist}{If TRUE and the output file already exists,
\code{JPlotToPDF} will do nothing.}

\item{createDirectory}{If TRUE and \code{filename} is located in a directory
which doesn't exist, the directory will be created.}

\item{...}{Any additional arguments are passed to
\code{\link[grDevices]{pdf}}, for example, text and font control parameters
such as \code{pointsize} and \code{family}.}
}
\value{
The result of evaluating \code{plotExpr} is returned invisibly (which
  means it is not automatically printed).
}
\description{
Writes the output of a plot to a PDF file. If you are using ggplot, try
either using \code{JPlotToPDF(filename, print(<plotting code>))} or else
\code{ggsave()}.
}
\details{
The \code{\link[grDevices]{pdf}} argument \code{pointsize} can be specified to control the resolution of the image (as can all other arguments to \code{pdf}.)
}
\seealso{
\code{\link[grDevices]{pdf}}
}
