% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_file.R
\name{JPlotToPNG}
\alias{JPlotToPNG}
\title{Plot to a PNG file}
\usage{
JPlotToPNG(
  filename,
  plotExpr,
  width = 180,
  height = NA,
  aspectRatio = 3/2,
  units = c("mm", "cm", "px", "in"),
  type = ifelse(capabilities()["cairo"], "cairo", NULL),
  res = 72,
  onlyIfDoesntExist = FALSE,
  createDirectory = TRUE,
  ...
)
}
\arguments{
\item{filename}{The name of the PNG to create or overwrite. If NULL, plot output goes to the current device.}

\item{plotExpr}{A function or expression which will produce the plot to be
written to the file.}

\item{width}{The width of the output PNG file in \code{units}.}

\item{height}{The height of the output PNG file in \code{units}. Defaults to
\code{width / aspectRatio}. If \code{height} is specified,
\code{aspectRatio} will be ignored.}

\item{aspectRatio}{Aspect ratio (\code{width / height}) of the output PNG
file.}

\item{units}{Units of \code{width} and \code{height}.}

\item{type}{Plotting device - defaults to "cairo" if that is an available
device since it produces nicer looking graphics.}

\item{res}{The nominal resolution in ppi. The value is passed in to
\code{\link[grDevices]{png}}. Increasing the resolution will increase the
size (in pixels) of the text and graph elements.}

\item{onlyIfDoesntExist}{If TRUE and the output file already exists,
\code{JPlotToPNG} will do nothing.}

\item{createDirectory}{If TRUE and \code{filename} is located in a directory
which doesn't exist, the directory will be created.}

\item{...}{Any additional arguments are passed to
\code{\link[grDevices]{png}}.}
}
\value{
The result of evaluating \code{plotExpr} is returned invisibly (which
  means it is not automatically printed).
}
\description{
Writes the output of a plot to a PNG file. If you are using ggplot, try
either using \code{JPlotToPNG(filename, print(<plotting code>))} or else
\code{ggsave()}.
}
\examples{
JPlotToPNG("test.png", plot(rnorm(50), type = 'l', main = 'Wiggly line'))

# Plot to a PNG file with width 180 mm, height 120 mm
# (i.e. height / aspectRatio which defaults to (3 / 2)), resolution 300 ppi.
# This results in a PNG file with size 2125x1417 pixels
JPlotToPNG("test.png", plot(1:10 + rnorm(10), type = "o"), width = 180, units = "mm", res = 300)

}
\seealso{
\code{\link[grDevices]{png}}
}
