% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_file.R
\name{JPlotToSVG}
\alias{JPlotToSVG}
\title{Plot to an SVG file}
\usage{
JPlotToSVG(filename, plotExpr, width = 180, height = NA,
  aspectRatio = 3/2, units = c("mm", "cm", "in"), bg = "white",
  family = "Helvetica", onlyIfDoesntExist = FALSE,
  createDirectory = TRUE, ...)
}
\arguments{
\item{filename}{The name of the SVG file to create or overwrite.}

\item{plotExpr}{A function or expression which will produce the plot to be
written to the file.}

\item{width}{The width of the output SVG file in \code{units}.}

\item{height}{The height of the output SVG file in \code{units}. Defaults to
\code{width / aspectRatio}. If \code{height} is specified,
\code{aspectRatio} will be ignored.}

\item{aspectRatio}{Aspect ratio (\code{width / height}) of the output EPS
file.}

\item{units}{Units of \code{width} and \code{height}.}

\item{bg}{Background colour - may be "transparent" for no background.}

\item{family}{The font family to be used. Passed to
\code{\link[grDevices]{svg}}.}

\item{onlyIfDoesntExist}{If TRUE and the output file already exists,
\code{JPlotToEPS} will do nothing.}

\item{createDirectory}{If TRUE and \code{filename} is located in a directory
which doesn't exist, the directory will be created.}

\item{...}{Any additional arguments are passed to
\code{\link[grDevices]{svg}}.}
}
\value{
The result of evaluating \code{plotExpr} is returned invisibly (which
  means it is not automatically printed).
}
\description{
Writes the output of a plot to a SVG file. If you are using ggplot, try
either using \code{JPlotToSVG(filename, print(<plotting code>))} or else
\code{ggsave()}.
}
\examples{
\dontrun{
plotWithAlpha <- function() {
    plot(1:3, type = 'l', lwd = 20, col = "#ffcc8888")
    lines(3:1, lwd = 20, col = "#88ccff88")
}

JPlotToSVG(filename, plotWithAlpha)
}

}
\seealso{
\code{\link[grDevices]{cairo}}
}
