% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_file.R
\name{JPlotToTIFF}
\alias{JPlotToTIFF}
\title{Plot to a TIFF file}
\usage{
JPlotToTIFF(
  filename,
  plotExpr,
  width = 180,
  height = NA,
  aspectRatio = 3/2,
  units = c("mm", "cm", "px", "in"),
  tryToUseRagg = TRUE,
  type = ifelse(capabilities()["cairo"], "cairo", NULL),
  res = 72,
  onlyIfDoesntExist = FALSE,
  createDirectory = TRUE,
  ...
)
}
\arguments{
\item{filename}{The name of the TIFF to create or overwrite. If NULL, plot output goes to the current device.}

\item{plotExpr}{A function or expression which will produce the plot to be
written to the file.}

\item{width}{The width of the output TIFF file in \code{units}.}

\item{height}{The height of the output TIFF file in \code{units}. Defaults to
\code{width / aspectRatio}. If \code{height} is specified,
\code{aspectRatio} will be ignored.}

\item{aspectRatio}{Aspect ratio (\code{width / height}) of the output TIFF
file.}

\item{units}{Units of \code{width} and \code{height}.}

\item{tryToUseRagg}{If \code{TRUE} (the default) and the \code{ragg} package
is installed, the TIFF file will be created using
\code{\link[ragg]{agg_tiff}}, otherwise \code{\link[grDevices]{tiff}} will be
used. See \url{https://ragg.r-lib.org} for more information.}

\item{type}{Plotting device; passed directly to \code{\link[grDevices]{tiff}}.
Defaults to "cairo" if that is an available device since it produces nicer
looking graphics. Ignored if \code{tryToUseRagg == TRUE} and \code{ragg} is
installed.}

\item{res}{The nominal resolution in ppi. The value is passed in to
\code{\link[grDevices:png]{grDevices::tiff()}}. Increasing the resolution
will increase the size (in pixels) of the text and graph elements.}

\item{onlyIfDoesntExist}{If TRUE and the output file already exists,
\code{JPlotToTIFF} will do nothing.}

\item{createDirectory}{If TRUE and \code{filename} is located in a directory
which doesn't exist, the directory will be created.}

\item{...}{Any additional arguments are passed to
\code{\link[ragg]{agg_tiff}} or \code{\link[grDevices]{tiff}}.}
}
\value{
The result of evaluating \code{plotExpr} is returned invisibly (which
  means it is not automatically printed).
}
\description{
Writes the output of a plot to a TIFF file. If you are using ggplot, try
either using \code{JPlotToTIFF(filename, print(<plotting code>))} or else
\code{ggsave()}.
}
\seealso{
\code{\link[ragg]{agg_tiff}}, \code{\link[grDevices]{tiff}}
}
