% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_file.R
\name{JReportToFile}
\alias{JReportToFile}
\title{Send console (i.e. text) output to a file}
\usage{
JReportToFile(filename, expr)
}
\arguments{
\item{filename}{Name of the file to write to.}

\item{expr}{An expression which outputs the text to be written.}
}
\description{
Sends the text output of an arbitrary R expression of function to a text file.
}
\examples{
JReportToFile("test.txt", print("Hello world!"))

}
