% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim-tools.R
\name{JScenesnFrames}
\alias{JScenesnFrames}
\title{Returns the total number of frames in a scene list.}
\usage{
JScenesnFrames(scenes)
}
\arguments{
\item{scenes}{A list of scene objects, each one created by calling the
\code{\link{JScene}} constructor.}
}
\value{
Number of frames in the animation specified by \code{scenes}.
}
\description{
Returns the total number of frames in a scene list.
}
