% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{JToSentence}
\alias{JToSentence}
\title{Convert a vector to a human readable list}
\usage{
JToSentence(v, sep = ", ", conjunction = " and ")
}
\arguments{
\item{v}{Vector of values.}

\item{sep}{Separator to use between the first elements of the list.}

\item{conjunction}{Value to use before the last element in the list.}
}
\description{
Similar to calling `paste(v, collapse = ", ")` except that the conjuction ("
and " by default) is used as the final separator.
}
\examples{
print(JToSentence(c("apple", "banana", "mandarin", "mango")))
# => [1] "apple, banana, mandarin and mango"

}
