\name{jutils-package}
\alias{jutils-package}
\alias{jutils}
\docType{package}
\title{
\packageTitle{jutils}
}
\description{
\packageDescription{jutils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{jutils}
\packageIndices{jutils}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{jutils}

Maintainer: \packageMaintainer{jutils}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
