% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{loadBase64RData}
\alias{loadBase64RData}
\title{Load the contents of an RData file that is stored in a Base64 string}
\usage{
loadBase64RData(.base64)
}
\arguments{
\item{.base64}{The Base64 string containing the RData file}
}
\value{
A list containing the contents of the file in .base64
}
\description{
Load the contents of an RData file that is stored in a Base64 string
}
