% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{loadRdata}
\alias{loadRdata}
\title{Load an RData file and return the contents, i.e. dont load it into the global scope}
\usage{
loadRdata(file_name)
}
\arguments{
\item{file_name}{The name of the file to load}
}
\value{
The contents of the file in file_name
}
\description{
Load an RData file and return the contents, i.e. dont load it into the global scope
}
