% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{monthDiff}
\alias{monthDiff}
\title{Get the number of months between two dates.}
\usage{
monthDiff(start, end)
}
\arguments{
\item{start}{The start date.}

\item{end}{The end date.}
}
\value{
An integer which is the number of (full) months between start and end.
}
\description{
Get the number of months between two dates.
}
