% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{monthEnd}
\alias{monthEnd}
\title{Get the last day of a month given a date.}
\usage{
monthEnd(x)
}
\arguments{
\item{x}{The date to get the last day of the month of.}
}
\value{
A Date which is the last day of the month of x.
}
\description{
Get the last day of a month given a date.
}
