% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{multiplot}
\alias{multiplot}
\title{Plot many columns of data in a matrix in the same plot}
\usage{
multiplot(
  mat,
  logscale = F,
  ylim = c(min(mat), max(mat)),
  legend = F,
  names = names(mat),
  ...
)
}
\arguments{
\item{mat}{The matrix containing the data to be plotted}

\item{logscale}{Should the data be log transformed before plotting}

\item{ylim}{The limits of the y axis of the plot, default to the min and max of mat}

\item{legend}{Should a legend be appended to the plot? Defaults to false}

\item{names}{The names to use in the legend, defaults to the names of the matrix mat}
}
\description{
Plot many columns of data in a matrix in the same plot
}
