% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{periodEnd}
\alias{periodEnd}
\title{Get the last day of a period given a date.}
\usage{
periodEnd(x, frequency, sunday = TRUE)
}
\arguments{
\item{x}{The date to get the last day of the period of.}

\item{frequency}{The frequency denoting the period, should be 1 (year), 4 (quarter), 12 (month) or 52 (week).}

\item{sunday}{Should weeks start on sunday, default = TRUE, otherwise weeks start on sunday. Only used if frequency == 52.}
}
\value{
A Date which is the last day of the period of x.
}
\description{
Get the last day of a period given a date.
}
