% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{progressbar}
\alias{progressbar}
\title{Print a progress bar while iterating over something}
\usage{
progressbar(progress, size = 40)
}
\arguments{
\item{progress}{A integer denoting how far we have come}

\item{size}{The total size of the progress bar}
}
\value{
progress + 1, to update the calling function
}
\description{
Print a progress bar while iterating over something
}
