% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{quantmean}
\alias{quantmean}
\title{Generic function for calculating quantiles and mean of an object}
\usage{
quantmean(x, ...)
}
\arguments{
\item{x}{The object to process}

\item{...}{Additional arguments}
}
\description{
Generic function for calculating quantiles and mean of an object
}
