% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{quarterEnd}
\alias{quarterEnd}
\title{Get the last day of a quarter given a date.}
\usage{
quarterEnd(x)
}
\arguments{
\item{x}{The date to get the last day of the quarter of.}
}
\value{
A Date which is the last day of the v of x.
}
\description{
Get the last day of a quarter given a date.
}
