% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{quarterStart}
\alias{quarterStart}
\title{Get the first day of a quarter given a date.}
\usage{
quarterStart(x)
}
\arguments{
\item{x}{The date to get the first day of the quarter of.}
}
\value{
A Date which is the first day of the quarter of x.
}
\description{
Get the first day of a quarter given a date.
}
