% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{weekEnd}
\alias{weekEnd}
\title{Get the last day of a week given a date.}
\usage{
weekEnd(x, saturday = TRUE)
}
\arguments{
\item{x}{The date to get the last day of the week of.}

\item{saturday}{Should weeks end on saturday, default = TRUE, otherwise weeks end on sunday.}
}
\value{
A Date which is the last day of the week of x.
}
\description{
Get the last day of a week given a date.
}
