% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{weekStart}
\alias{weekStart}
\title{Get the first day of a week given a date.}
\usage{
weekStart(x, sunday = TRUE)
}
\arguments{
\item{x}{The date to get the first day of the week of.}

\item{sunday}{Should weeks start on sunday, default = TRUE, otherwise weeks start on monday.}
}
\value{
A Date which is the first day of the week of x.
}
\description{
Get the first day of a week given a date.
}
