% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{yearEnd}
\alias{yearEnd}
\title{Get the last day of a year given a date.}
\usage{
yearEnd(x)
}
\arguments{
\item{x}{The date to get the last day of the year of.}
}
\value{
A Date which is the last day of the year of x.
}
\description{
Get the last day of a year given a date.
}
