% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{yearStart}
\alias{yearStart}
\title{Get the first day of a year given a date.}
\usage{
yearStart(x)
}
\arguments{
\item{x}{The date to get the first day of the year of.}
}
\value{
A Date which is the first day of the year of x.
}
\description{
Get the first day of a year given a date.
}
