% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ciPlot}
\alias{ciPlot}
\title{Create a confidence interval plot}
\usage{
ciPlot(
  data,
  col = 1,
  append = FALSE,
  ylim = c(min(as.matrix(data$low)[, col]), max(as.matrix(data$high)[, col])),
  ci.col = "lightgrey",
  ci.density = 100,
  ci.angle = 0,
  ci.border = NA,
  mean = TRUE,
  mean.lty = "solid",
  median = FALSE,
  median.lty = "solid",
  ...
)
}
\arguments{
\item{data}{The data to plot, should be a list that contains "mean", "low" and "high" matrices}

\item{col}{Which column of each matrix contains the data to plot, defaults to 1}

\item{append}{Should we create a new plot or append a confidence interval to one already active, defaults to FALSE}

\item{ylim}{The y axis limits for the plot, defaults to min of low and max of high}

\item{ci.col}{The confidence area color, defaults to lightgrey}

\item{ci.density}{The confidence area density, defaults to 100}

\item{ci.angle}{The confidence area angle, defaults to 0}

\item{ci.border}{The confidence area border, defaults to NA}

\item{mean}{Should the mean be plotted, default is TRUE}

\item{mean.lty}{Line type of mean, default is solid}

\item{median}{Should median be plotted, default is FALSE}

\item{median.lty}{Line type of median, default is solid}

\item{...}{Additional arguments to pass to the plot function.}
}
\description{
Create a confidence interval plot
}
