% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim-tools.R
\name{JAnimateScenes}
\alias{JAnimateScenes}
\title{Animate a list of \code{\link{JScene}}s.}
\usage{
JAnimateScenes(videoFileName, scenes, ...)
}
\arguments{
\item{videoFileName}{Name of the GIF file to be created.}

\item{scenes}{A list of scene objects, each one created by calling the
\code{\link{JScene}} constructor.}

\item{...}{Additional parameters are passed on to \code{\link{JAnimateGIF}}.}
}
\value{
The list of arguments passed to the \code{\link{JAnimateGIF}}
  function (invisibly).
}
\description{
Animates a \code{list} of \code{\link{JScene}} objects by plotting each frame
and then combining them into a single animated GIF.
}
\details{
See \code{\link{JAnimateGIF}} for general information about creating
animations from R.
}
\examples{
\dontrun{
scenes <- list(JScene(1, 20,
                      pt1 = JTransition(1, 0, JEaseInOut),
                      pt2 = JTransition(0, 1),
                      plotFn = function(pt1, pt2) {
                        plot(c(pt1, pt2), c(1, pt2), type = "b", xlim = c(0, 1), ylim = c(0, 1))
                      }),

               JScene(1, 20,
                      pt1 = JTransition(0, 1, JEaseInOut),
                      pt2 = JTransition(1, 0, JEaseIn),
                      plotFn = function(pt1, pt2) {
                        plot(c(pt1, pt2), c(1, pt2), type = "b", xlim = c(0, 1), ylim = c(0, 1))
                      })
)
JAnimateScenes("animation.gif", scenes)
}

}
\seealso{
\code{\link{JAnimateGIF}}, \code{\link{JScene}}, \code{\link{JTransition}}, \code{\link{JEase}}, \code{\link{JPlotSceneFn}}
}
