% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim-tools.R
\name{JEase}
\alias{JEase}
\alias{JLinear}
\alias{JEaseIn}
\alias{JEaseOut}
\alias{JEaseInOut}
\alias{JBounce}
\title{Predefined transition timings}
\usage{
JEase(x)

JLinear(x)

JEaseIn(x)

JEaseOut(x)

JEaseInOut(x)

JBounce(x)
}
\arguments{
\item{x}{x-value passed to the function automatically. See \code{Examples} below.}
}
\value{
A timing function.
}
\description{
These functions are passed by name (i.e. without following parentheses) as the
\code{timing} parameter to the \code{\link{JTransition}} function.
}
\examples{
width = JTransition(.1, 1, JEaseInOut)

}
\seealso{
\code{\link{JTransition}}, \code{\link{JScene}}, \code{\link{JBezier}}
}
