% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim-tools.R
\name{JEase}
\alias{JEase}
\alias{JLinear}
\alias{JEaseIn}
\alias{JEaseOut}
\alias{JEaseInOut}
\alias{JBounce}
\title{Predefined transition timings}
\usage{
JEase(x)

JLinear(x)

JEaseIn(x)

JEaseOut(x)

JEaseInOut(x)

JBounce(x)
}
\arguments{
\item{x}{x-value passed to the function automatically. See \code{Examples} below.}
}
\value{
A timing function.
}
\description{
These functions are passed by name (i.e. without following parentheses) as the
\code{timing} parameter to the \code{\link{JTransition}} function.
}
\examples{
# Vary width from 0.1 to 1, with faster change in the middle of the scene
width = JTransition(0.1, 1, JEaseInOut)

# Function to graphically visualise a timing function
VisualiseTimingFun <- function(fn) plot(seq(0, 1, .01), fn(seq(0, 1, .01)), type = "l")
# Visualise the JEase timing function
VisualiseTimingFun(JEase)

}
\seealso{
\code{\link{JTransition}}, \code{\link{JScene}}, \code{\link{JBezier}}, \code{\link{JLoop}}
}
