% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_densities.R
\name{JPlotDensities}
\alias{JPlotDensities}
\title{Plot a list of densities}
\usage{
JPlotDensities(
  densities,
  lineColours = NULL,
  fillColours = NULL,
  fillAlpha = NA,
  fillDensities = NA,
  fillAngles = NA,
  lty = 1,
  lwd = 2,
  xlim = NULL,
  ylim = NULL,
  ylab = "Density",
  add = FALSE,
  legendLabels,
  legendPos = "topleft",
  ...
)
}
\arguments{
\item{densities}{List of densities to be plotted}

\item{lineColours}{Vector of colours used to draw the lines. Defaults to the
rainbow palette (\code{\link[grDevices]{rainbow}}).}

\item{fillColours}{Vector of colours to fill the density polygons. If not
specified or NULL, the polygons are not filled.}

\item{fillAlpha}{If not NA (the default), should be an alpha value between 0
and 1. The alpha of the fill colours are multiplied by this value.}

\item{fillDensities}{If not NA (the default), density of shading lines used
to fill polygons (passed as argument \code{density} to
\code{\link[graphics]{polygon}}).}

\item{fillAngles}{If not NA (the default), slope of shading lines used to
fill polygons (passed as argument \code{angle} to
\code{\link[graphics]{polygon}}).}

\item{lty}{Line style used to draw the lines.}

\item{lwd}{Line width used to draw the lines.}

\item{xlim}{Defines the graphical extents of the x-axis. Defaults to include
all the density lines.}

\item{ylim}{Defines the graphical extents of the y-axis. Defaults to include
all the density lines.}

\item{ylab}{Y label, passed to \code{\link[graphics]{plot}}.}

\item{add}{If TRUE, the densities are added to an existing plot, otherwise a
new plot is created.}

\item{legendLabels}{Optional vector of characters. If specified, a legend will be added to the plot.}

\item{legendPos}{Position of optional legend.}

\item{...}{Additional parameters are passed to \code{\link[graphics]{plot}}.}
}
\description{
Plots a list of densities (as created by \code{\link[stats]{density}}).
Plotting is performed using base graphics. Each density is drawn as a line,
and optionally filled. Fill colours may be easily made transparent by
specifying a value for \code{fillAlpha}. Attempts to provide sensible
defaults while allowing maximum flexibility.
}
\examples{
data <- list(
  normal = rnorm(100),
  uniform = runif(50),
  exponential = rexp(200))
densities <- lapply(data, density)
JPlotDensities(densities)
# Display legend
legend("topleft", c("Normal", "Uniform", "Exponential"), col = rainbow(3), lwd = 2)

}
\seealso{
\code{\link[stats]{density}}, \code{\link[grDevices]{rainbow}},
  \code{RColorBrewer::brewer.pal}, \code{\link[graphics]{polygon}}
}
