% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_file.R
\name{JPlotToEPS}
\alias{JPlotToEPS}
\title{Plot to an EPS (encapsulated postscript) file}
\usage{
JPlotToEPS(filename, plotExpr, width = 180, height = NA,
  aspectRatio = 3/2, units = c("mm", "cm", "in"), bg = "white",
  paper = "special", family = "Helvetica", onlyIfDoesntExist = F, ...)
}
\arguments{
\item{filename}{The name of the EPS to create or overwrite.}

\item{plotExpr}{A function or expression which will produce the plot to be
written to the file.}

\item{width}{The width of the output EPS file in \code{units}.}

\item{height}{The height of the output EPS file in \code{units}. Defaults to
\code{width / aspectRatio}. If \code{height} is specified,
\code{aspectRatio} will be ignored.}

\item{aspectRatio}{Aspect ratio (\code{width / height}) of the output EPS
file.}

\item{units}{Units of \code{width} and \code{height}.}

\item{bg}{Background colour - may be "transparent" for no background.}

\item{paper}{Paper size, defaults to "special" which is required if
\code{width} and \code{height} are to be used.}

\item{family}{The font family to be used.}

\item{onlyIfDoesntExist}{If TRUE and the output file already exists,
\code{JPlotToEPS} will do nothing.}

\item{...}{Any additional arguments are passed to
\code{\link[grDevices]{pdf}}.}
}
\description{
Writes the output of a plot to a EPS file. If you are using ggplot, try
either using \code{JPlotToEPS(filename, print(<plotting code>))} or else
\code{ggsave()}.
}
\seealso{
\code{\link[grDevices]{setEPS}}, \code{\link[grDevices]{postscript}}
}
