% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_file.R
\name{JPlotToFile}
\alias{JPlotToFile}
\title{Plot to a file}
\usage{
JPlotToFile(filename, plot, ...)
}
\arguments{
\item{filename}{The name of the file to create or overwrite.}

\item{plot}{A function or expression which will produce the plot to be
written to the file.}

\item{...}{Any additional arguments are passed to the appropriate function.}
}
\description{
Writes the output of a plot to a file. The type of file is deduced from the
extension of the file name. If you are using ggplot rather than base
graphics, try either using \code{JPlotToFile(filename, print(<plotting
code>))} or \code{ggsave()}.
}
\seealso{
\code{\link{JPlotToPNG}}, \code{\link{JPlotToTIFF}},
  \code{\link{JPlotToPDF}}, \code{\link{JPlotToEPS}}
}
