% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_file.R
\name{JPlotToPDF}
\alias{JPlotToPDF}
\title{Plot to a PDF file}
\usage{
JPlotToPDF(
  filename,
  plotExpr,
  width = 180,
  height = NA,
  aspectRatio = 3/2,
  units = c("mm", "cm", "in"),
  bg = "white",
  paper = "special",
  family = "Helvetica",
  onlyIfDoesntExist = FALSE,
  createDirectory = TRUE,
  cairo = FALSE,
  embedFonts = FALSE,
  ...
)
}
\arguments{
\item{filename}{The name of the PDF to create or overwrite. If NULL, plot
output goes to the current device.}

\item{plotExpr}{A function or expression which will produce the plot to be
written to the file.}

\item{width}{The width of the output PDF file in \code{units}.}

\item{height}{The height of the output PDF file in \code{units}. Defaults to
\code{width / aspectRatio}. If \code{height} is specified,
\code{aspectRatio} will be ignored.}

\item{aspectRatio}{Aspect ratio (\code{width / height}) of the output PDF
file.}

\item{units}{Units of \code{width} and \code{height}.}

\item{bg}{Background colour - may be "transparent" for no background.}

\item{paper}{Paper size, defaults to "special" which is the value it must
have if \code{width} and \code{height} are to be used. Ignored if
\code{cairo} is \code{TRUE}.}

\item{family}{The font family to be used. Consider using \code{"sans"},
\code{"serif"} and \code{"mono"} for consistent viewing results across
devices.}

\item{onlyIfDoesntExist}{If TRUE and the output file already exists,
\code{JPlotToPDF} will do nothing.}

\item{createDirectory}{If TRUE and \code{filename} is located in a directory
which doesn't exist, the directory will be created.}

\item{cairo}{If TRUE, the \code{\link[grDevices]{cairo_pdf}} device is used,
otherwise the \code{\link[grDevices]{pdf}} device is used (see Details).}

\item{embedFonts}{If TRUE, \code{\link[grDevices]{embedFonts}} is called on
the file after creation. This is provided as a convenience. Ghostscript
must be installed and locatable (see \code{\link[grDevices]{embedFonts}}).
I \emph{think} this is unnecessary when using the Cairo PDF device.}

\item{...}{Any additional arguments are passed to
\code{\link[grDevices]{pdf}}, for example, text and font control parameters
such as \code{pointsize}.}
}
\value{
The result of evaluating \code{plotExpr} is returned invisibly (which
  means it is not automatically printed).
}
\description{
Writes the output of a plot to a PDF file. If you are using ggplot, try
either using \code{JPlotToPDF(filename, print(<plotting code>))} or else
\code{ggsave()}.
}
\details{
The \code{\link[grDevices]{pdf}} argument \code{pointsize} can be specified
to control the resolution of the image (as can all other arguments to
\code{pdf}.)

The pdf device does \emph{not} embed fonts in the pdf file. This probably
does not matter if you use one of the device-independent font families,
\code{"sans"}, \code{"serif"} and \code{"mono"}, because an appropriate (but
not necessarily identical) font will be used by the viewer. Alternatively,
specifying \code{cairo = TRUE} \emph{may} embed the font, although the
documentation for \code{\link[grDevices]{cairo_pdf}} is confusing.
}
\seealso{
\code{\link[grDevices]{pdf}}, \code{\link[grDevices]{cairo_pdf}},
  \code{\link[grDevices]{embedFonts}}
}
