% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_file.R
\name{JPlotToTIFF}
\alias{JPlotToTIFF}
\title{Plot to a TIFF file}
\usage{
JPlotToTIFF(filename, plotExpr, width = 180, height = NA,
  aspectRatio = 3/2, units = c("mm", "cm", "px", "in"),
  type = ifelse(capabilities()["cairo"], "cairo", NULL), res = 72,
  onlyIfDoesntExist = F, ...)
}
\arguments{
\item{filename}{The name of the TIFF to create or overwrite.}

\item{plotExpr}{A function or expression which will produce the plot to be
written to the file.}

\item{width}{The width of the output TIFF file in \code{units}.}

\item{height}{The height of the output TIFF file in \code{units}. Defaults to
\code{width / aspectRatio}. If \code{height} is specified,
\code{aspectRatio} will be ignored.}

\item{aspectRatio}{Aspect ratio (\code{width / height}) of the output TIFF
file.}

\item{units}{Units of \code{width} and \code{height}.}

\item{type}{Plotting device - defaults to "cairo" if that is an available
device since it produces nicer looking graphics.}

\item{res}{The nominal resolution in ppi. The value is passed in to
\code{\link[grDevices]{tiff}}. Increasing the resolution will increase the
size (in pixels) of the text and graph elements.}

\item{onlyIfDoesntExist}{If TRUE and the output file already exists,
\code{JPlotToTIFF} will do nothing.}

\item{...}{Any additional arguments are passed to
\code{\link[grDevices]{tiff}}.}
}
\description{
Writes the output of a plot to a TIFF file. If you are using ggplot, try either
using \code{JPlotToTIFF(filename, print(<plotting code>))} or else \code{ggsave()}.
}
\seealso{
\code{\link[grDevices]{tiff}}
}
