% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim-tools.R
\name{JScene}
\alias{JScene}
\title{Construct a JScene, which is a portion of a complex animation. A JScene
consists of some metadata (duration, frame rate), a set of animation
transitions and a plotting function.}
\usage{
JScene(duration, fps, startAfter = 0, ..., plotFn)
}
\arguments{
\item{duration}{Scene duration in seconds.}

\item{fps}{Frame rate (frames per second).}

\item{startAfter}{Timing of the start of this frame. Number of seconds after
the end of the previous scene that this scene starts.}

\item{...}{Ordered set of transitions. Transitions may be named for
documentation purposes, but the names are not used.}

\item{plotFn}{The parameterised plotting function. Called with one positional
argument for each transition, and an optional final boolean argument,
\code{add}, which is \code{TRUE} if the function should add to an existing
plot rather than create a new plot. The \code{add} is only used for
overlapping scenes.}
}
\value{
A list, known as a _JScene_, which can be included in a list of
  scenes then used to create an animation.
}
\description{
Construct a JScene, which is a portion of a complex animation. A JScene
consists of some metadata (duration, frame rate), a set of animation
transitions and a plotting function.
}
\seealso{
\code{\link{JAnimateScenes}}
}
