\name{jutils-package}
\alias{jutils-package}
\alias{jutils}
\docType{package}
\title{
\packageTitle{jutils}
}
\description{
\packageDescription{jutils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{jutils}
\packageIndices{jutils}

}
\author{
\packageAuthor{jutils}

Maintainer: \packageMaintainer{jutils}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
# simple examples of the most important functions
}
