% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{quantmean.list}
\alias{quantmean.list}
\title{Function for calculating the quantiles and mean of many matrices in a list}
\usage{
\method{quantmean}{list}(matlist, quantiles = c(0.05, 0.95), mean = TRUE)
}
\arguments{
\item{matlist}{The list of matrices to use}

\item{quantiles}{The quantiles to calculate, defaults to 0.05 and 0.95 (i.e. 90\%)}

\item{mean}{Should the mean also be calculated? Defaults to TRUE}
}
\value{
A list containing the mean per element over all the matrices, together with the low and high quantiles selected
}
\description{
Function for calculating the quantiles and mean of many matrices in a list
}
